requirements.txt
yt-dlp>=2024.4.9
openai-whisper>=20231117
transformers>=4.40.0
torch>=2.1.0
torchaudio>=2.1.0
numpy>=1.26.0
accelerate>=0.30.0
Flask
Flask-Cors
reportlab
python-docx